+++ mongoose/src/decorators/schema.ts
@@ -54,19 +54,3 @@ export function Index() {
     getMongooseMeta(target).indexes.push(name);
   };
 }
-
-/**
- * Defines set method - options for model
- */
-export function Set() {
-  console.warn(`Deprecated: This decorator will be removed in a future release.
-  Use the options parameter of the Model decorator instead.`);
-  return (target: any, name: string) => {
-    getMongooseMeta(target).options.push(name);
-  };
-}
-/**
- * Alias of Set
- * @see {Set}
- */
-export const Option = Set;