+++ playground/mongoose/index.ts
@@ -1,14 +1,18 @@
-import { TestModel } from './model';
+import { Animal, AnimalSchema } from './model';
 
+console.log(AnimalSchema.create);
 
-TestModel.staticMethod();
-let test = new TestModel({testField: "Hello World"});
+Animal.staticMethod();
+console.log(Animal.staticField);
+
+let test = new Animal({ testField: 'Hello World' });
+
 console.log('Model = ' + test.toString());
+
 test.setField();
+
+test.testField = 'Test Field';
 
 test.instanceMethod();
 
-console.log('Model = ' + test.testField);
-
-process.exit();
-
+process.exit(0);